+++ src/IpAddressVector.cpp
@@ -544,3 +544,60 @@ LogicalVector IpAddressVector::isIPv4Mapped() const {
 
   return output;
 }
+
+
+/*-----------------------*
+ *  Address translation  *
+ * ----------------------*/
+IpAddressVector IpAddressVector::mapToIPv4(bool keep_ipv6) const {
+  std::size_t vsize = is_na.size();
+
+  // initialize vectors
+  std::vector<asio::ip::address_v4> out_address_v4(vsize);
+  std::vector<asio::ip::address_v6> out_address_v6(vsize);
+  std::vector<bool> out_is_ipv6(vsize, false);
+  std::vector<bool> out_is_na(vsize, false);
+
+  for (std::size_t i=0; i<vsize; ++i) {
+    if (is_na[i]) {
+      out_is_na[i] = true;
+    } else if (is_ipv6[i]) {
+      if (address_v6[i].is_v4_mapped()) {
+        out_address_v4[i] = asio::ip::make_address_v4(asio::ip::v4_mapped, address_v6[i]);
+      } else if (keep_ipv6) {
+        out_address_v6[i] = address_v6[i];
+        out_is_ipv6[i] = true;
+      } else {
+        out_is_na[i] = true;
+      }
+    } else {
+      out_address_v4[i] = address_v4[i];
+    }
+  }
+
+  return IpAddressVector(out_address_v4, out_address_v6, out_is_ipv6, out_is_na);
+}
+
+IpAddressVector IpAddressVector::mapToIPv6() const {
+  std::size_t vsize = is_na.size();
+
+  // initialize vectors
+  std::vector<asio::ip::address_v4> out_address_v4(vsize);
+  std::vector<asio::ip::address_v6> out_address_v6(vsize);
+  std::vector<bool> out_is_ipv6(vsize, false);
+  std::vector<bool> out_is_na(vsize, false);
+
+  for (std::size_t i=0; i<vsize; ++i) {
+    if (is_na[i]) {
+      out_is_na[i] = true;
+    } else if (is_ipv6[i]) {
+      out_address_v6[i] = address_v6[i];
+      out_is_ipv6[i] = true;
+    } else {
+      out_address_v6[i] = asio::ip::make_address_v6(asio::ip::v4_mapped, address_v4[i]);
+      out_is_ipv6[i] = true;
+    }
+  }
+
+  return IpAddressVector(out_address_v4, out_address_v6, out_is_ipv6, out_is_na);
+}