+++ NEWS.md
@@ -5,6 +5,7 @@
 * New `seq.ip_network()` and `hosts()` functions to list all addresses within a network.
 * New `sample_ip()` function to randomly sample addresses within a network.
 * New `packed()` and `unpack()` functions to encode and decode addresses as raw bytes.
+* New `ipv4_mapped()` and `ipv6_mapped()` functions map between address spaces (where possible).
 * New functions to provide basic information about a network: `num_addresses()`, `network_address()` and `broadcast_address()`.
 * New functions to check if addresses or networks are reserved for special use: `is_multicast()`, `is_unspecified()`, `is_loopback()`, `is_link_local()`, `is_ipv4_mapped()`.
 * New `max_prefix_length()` function to get the size of the address space (32-bit for IPv4 and 128-bit for IPv6).